backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.0.3
httpx==0.25.2

# PDF processing and text extraction
pdfplumber==0.10.3
python-docx==1.1.0
langchain-text-splitters==0.2.4

# Embeddings and vector storage
openai==1.6.1
numpy==1.24.3
scipy==1.11.4
faiss-cpu==1.7.4

# Alternative AI APIs
anthropic==0.8.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
isort==5.12.0
flake8==6.1.0